Use gitignore template
This repo has `node_modules` in it, which are all the dependencies stored locally. This can make the repository really big and hard to clone. Adding the directory to .gitignore means changes to them won't be recorded in the future. They're still in the git history, however, so committing the deletion of them won't make the repository itself smaller since it's still part of history. To delete history, check this out: https://gist.github.com/paulallies/0052fab554b14bbfa3ef.